+++ src/app/Components/ComponentsSamples/Tags/TagsSamples.xaml.cs
@@ -1,0 +1,9 @@
+namespace Components.ComponentsSamples.Tags;
+
+public partial class TagsSamples
+{
+    public TagsSamples()
+    {
+        InitializeComponent();
+    }
+}