+++ src/library/DIPS.Mobile.UI/Components/Alerting/SystemMessage/SystemMessageConfigurator.cs
@@ -6,9 +6,9 @@ namespace DIPS.Mobile.UI.Components.Alerting.SystemMessage;
 
 internal class SystemMessageConfigurator : ISystemMessageConfigurator
 {
-    public Color BackgroundColor { get; set; } = Colors.GetColor(ColorName.color_neutral_70);
-    public Color TextColor { get; set; } = Colors.GetColor(ColorName.color_text_on_action);
-    public Color IconColor { get; set; } = Colors.GetColor(ColorName.color_icon_on_action);
+    public Color BackgroundColor { get; set; } = Colors.GetColor(ColorName.color_surface_subtle);
+    public Color TextColor { get; set; } = Colors.GetColor(ColorName.color_text_action);
+    public Color IconColor { get; set; } = Colors.GetColor(ColorName.color_icon_action);
     public float Duration { get; set; } = 2500;
     public string Text { get; set; } = DUILocalizedStrings.YouHaveNotSetSomeText;
     [TypeConverter(nameof(ImageSourceConverter))]