+++ src/library/DIPS.Mobile.UI/Components/Buttons/Android/ButtonHandler.cs
@@ -51,7 +51,7 @@ public partial class ButtonHandler : Microsoft.Maui.Handlers.ButtonHandler
     {
         try
         {
-            var rippleColor = Resources.Colors.Colors.GetColor(ColorName.color_icon_neutral);
+            var rippleColor = Resources.Colors.Colors.GetColor(ColorName.color_icon_action);
             var ripple = new RippleDrawable(new Color(rippleColor.Red, rippleColor.Green, rippleColor.Blue, .2f).ToDefaultColorStateList(),
                 null,
                 PlatformView.Background);