+++ src/library/DIPS.Mobile.UI/Components/Selection/iOS/SwitchHandler.cs
@@ -14,6 +14,6 @@ public class SwitchHandler : Microsoft.Maui.Handlers.SwitchHandler
         // https://stackoverflow.com/questions/10348869/change-color-of-uiswitch-in-off-state
         platformView.Layer.CornerRadius = platformView.Frame.Height / 2;
         platformView.ClipsToBounds = true;
-        platformView.BackgroundColor = Colors.GetColor(ColorName.color_stroke_default).ToPlatform();
+        platformView.BackgroundColor = Colors.GetColor(ColorName.color_border_default).ToPlatform();
     }
 }