+++ src/library/DIPS.Mobile.UI/Resources/Colors/ColorResourcesDark.cs
@@ -3,7 +3,7 @@ Do not edit directly,
 this file is generated
 */
 
-namespace DIPS.Mobile.UI.Resources.Colors.Experimental;
+namespace DIPS.Mobile.UI.Resources.Colors;
 
 internal static class ColorResourcesDark
 {
@@ -25,6 +25,7 @@ internal static class ColorResourcesDark
         ["color_border_input_readonly"] = Color.FromRgba("#8f9496ff"), // neutral-500
         ["color_border_link"] = Color.FromRgba("#c0cff0ff"), // navy-300
         ["color_border_subtle"] = Color.FromRgba("#00107eff"), // navy-800
+        ["color_border_success"] = Color.FromRgba("#ddf5d8ff"), // green-200
         ["color_data_data_1"] = Color.FromRgba("#8da9e4ff"), // navy-400
         ["color_data_data_1_border"] = Color.FromRgba("#8da9e4ff"), // navy-400
         ["color_data_data_1_subtle"] = Color.FromRgba("#00107eff"), // navy-800
@@ -45,8 +46,8 @@ internal static class ColorResourcesDark
         ["color_fill_action_secondary_active"] = Color.FromRgba("#0000ccff"), // navy-700
         ["color_fill_action_secondary_hover"] = Color.FromRgba("#00107eff"), // navy-800
         ["color_fill_danger"] = Color.FromRgba("#f3583dff"), // red-700
-        ["color_fill_danger_active"] = Color.FromRgba("#ab321dff"), // red-800
-        ["color_fill_danger_hover"] = Color.FromRgba("#ab321dff"), // red-800
+        ["color_fill_danger_active"] = Color.FromRgba("#f9a698ff"), // red-500
+        ["color_fill_danger_hover"] = Color.FromRgba("#f98572ff"), // red-600
         ["color_fill_danger_subtle"] = Color.FromRgba("#ab321dff"), // red-800
         ["color_fill_default"] = Color.FromRgba("#00107eff"), // navy-800
         ["color_fill_default_active"] = Color.FromRgba("#c0cff0ff"), // navy-300
@@ -71,6 +72,7 @@ internal static class ColorResourcesDark
         ["color_fill_warning_hover"] = Color.FromRgba("#ecb62dff"), // yellow-800
         ["color_icon_action"] = Color.FromRgba("#c0cff0ff"), // navy-300
         ["color_icon_action_hover"] = Color.FromRgba("#e1e9faff"), // navy-200
+        ["color_icon_action_on_fill"] = Color.FromRgba("#ffffffff"), // neutral-white
         ["color_icon_danger"] = Color.FromRgba("#f9eae2ff"), // red-200
         ["color_icon_default"] = Color.FromRgba("#ffffffff"), // neutral-white
         ["color_icon_disabled"] = Color.FromRgba("#b6b7b9ff"), // neutral-400
@@ -83,6 +85,7 @@ internal static class ColorResourcesDark
         ["color_icon_on_fill_default_focus"] = Color.FromRgba("#ffffffff"), // neutral-white
         ["color_icon_on_fill_default_hover"] = Color.FromRgba("#ffffffff"), // neutral-white
         ["color_icon_on_fill_default_selected"] = Color.FromRgba("#000000ff"), // neutral-black
+        ["color_icon_on_fill_inverted"] = Color.FromRgba("#ffffffff"), // neutral-white
         ["color_icon_on_fill_success_hover"] = Color.FromRgba("#0f7157ff"), // green-800
         ["color_icon_readonly"] = Color.FromRgba("#8f9496ff"), // neutral-500
         ["color_icon_subtle"] = Color.FromRgba("#dadbdcff"), // neutral-300
@@ -92,17 +95,19 @@ internal static class ColorResourcesDark
         ["color_surface_default"] = Color.FromRgba("#050948ff"), // navy-900
         ["color_surface_information"] = Color.FromRgba("#095abfff"), // blue-800
         ["color_surface_neutral"] = Color.FromRgba("#373d40ff"), // neutral-800
-        ["color_surface_subtle"] = Color.FromRgba("#00107eff"), // navy-800
+        ["color_surface_subtle"] = Color.FromRgba("#0000ccff"), // navy-700
         ["color_surface_success"] = Color.FromRgba("#0f7157ff"), // green-800
-        ["color_surface_visited"] = Color.FromRgba("#0000ccff"), // navy-700
+        ["color_surface_visited"] = Color.FromRgba("#00107eff"), // navy-800
         ["color_surface_warning"] = Color.FromRgba("#ecb62dff"), // yellow-800
         ["color_text_action"] = Color.FromRgba("#c0cff0ff"), // navy-300
         ["color_text_action_hover"] = Color.FromRgba("#e1e9faff"), // navy-200
         ["color_text_action_neutral"] = Color.FromRgba("#ffffffff"), // neutral-white
+        ["color_text_action_on_fill"] = Color.FromRgba("#ffffffff"), // neutral-white
         ["color_text_danger"] = Color.FromRgba("#f9eae2ff"), // red-200
         ["color_text_default"] = Color.FromRgba("#ffffffff"), // neutral-white
         ["color_text_default_inverted"] = Color.FromRgba("#000000ff"), // neutral-black
         ["color_text_disabled"] = Color.FromRgba("#b6b7b9ff"), // neutral-400
+        ["color_text_error"] = Color.FromRgba("#f9eae2ff"), // red-200
         ["color_text_information"] = Color.FromRgba("#d8efffff"), // blue-200
         ["color_text_input"] = Color.FromRgba("#ffffffff"), // neutral-white
         ["color_text_label"] = Color.FromRgba("#ffffffff"), // neutral-white